requirements.txt
# --------- pytorch --------- #
torch==1.8.1
torchvision==0.9.1
pytorch-lightning>=1.3.2 #'1.3.8'
torchmetrics==0.3.2
timm
numpy==1.19.0

# --------- hydra --------- #
hydra-core==1.1.0.rc1
hydra-colorlog==1.1.0.dev1
hydra-optuna-sweeper==1.1.0.dev2
hydra-submitit-launcher 
# hydra-ax-sweeper==1.1.0
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.30
neptune-client==0.9.16
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit
black
isort
flake8

# --------- others --------- #
jupyterlab
python-dotenv
tqdm
rich
pytest
pandas
sh
scikit-learn
scikit-image==0.18.1
memory-profiler
seaborn
# dvc